requirements.txt
pytest
paramiko
pywinrm